refactor: 패키지 구조 리팩토링 및 패키지명 변경
