+++ src/main/java/com/gloddy/server/estimate/service/praise/PraiseService.java
@@ -1,7 +1,13 @@
 package com.gloddy.server.estimate.service;
 
+import ch.qos.logback.core.joran.action.AppenderRefAction;
 import com.gloddy.server.auth.entity.User;
-import com.gloddy.server.auth.service.UserFindService;
+import com.gloddy.server.core.error.handler.errorCode.ErrorCode;
+import com.gloddy.server.core.error.handler.exception.PraiseBusinessException;
+import com.gloddy.server.estimate.dto.PraiseResponse;
+import com.gloddy.server.estimate.entity.Praise;
+import com.gloddy.server.estimate.repository.PraiseJpaRepository;
+import com.gloddy.server.user.service.UserFindService;
 import com.gloddy.server.domain.AbsenceInGroupDomain;
 import com.gloddy.server.domain.UserPraise;
 import com.gloddy.server.estimate.dto.PraiseDto;
@@ -14,6 +20,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.gloddy.server.estimate.dto.PraiseResponse.*;
+
 @Service
 @RequiredArgsConstructor
 public class PraiseService {
@@ -21,6 +29,7 @@ public class PraiseService {
     private final GroupUserCountService groupUserCountService;
     private final AbsenceInGroupFindService absenceInGroupFindService;
     private final UserFindService userFindService;
+    private final PraiseJpaRepository praiseJpaRepository;
 
     @Transactional
     public void praiseInGroup(List<PraiseDto> praiseDtos, Long groupId) {
@@ -37,5 +46,19 @@ public class PraiseService {
                 .collect(Collectors.toUnmodifiableList());
 
         userPraises.forEach(UserPraise::applyPraisePoint);
+    }
+
+    @Transactional(readOnly = true)
+    public getPraiseForUser getPraiseForUser(Long userId) {
+        Praise praise = praiseJpaRepository.findByUserId(userId)
+                .orElseThrow(() -> new PraiseBusinessException(ErrorCode.PRAISE_NOT_FOUND));
+
+        return new getPraiseForUser(
+                praise.getTotalCalmCount(),
+                praise.getTotalKindCount(),
+                praise.getTotalActiveCount(),
+                praise.getTotalHumorCount(),
+                praise.getTotalAbsenceCount()
+        );
     }
 }
