+++ src/main/java/com/gloddy/server/user/handler/UserQueryHandlerImpl.java
@@ -1,7 +1,7 @@
-package com.gloddy.server.auth.handler;
+package com.gloddy.server.user.handler;
 
 import com.gloddy.server.auth.entity.User;
-import com.gloddy.server.auth.repository.UserRepository;
+import com.gloddy.server.user.repository.UserRepository;
 import com.gloddy.server.core.error.handler.errorCode.ErrorCode;
 import com.gloddy.server.core.error.handler.exception.UserBusinessException;
 import lombok.RequiredArgsConstructor;
